+++ src/main/java/com/springweb/domain/card/CardEntity.java
@@ -13,7 +13,7 @@ public class CardEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int cardCode; //카드 코드
+    private Long cardCode; //카드 코드
 
     @Column (nullable = false)
     private String cardName; //카드 명
@@ -58,7 +58,7 @@ public class CardEntity {
     private int count; //순위에 쓰일 카운트
 
     @Builder
-    public CardEntity(int cardCode, String cardName, String cardPhoto, String cardCompany, int annualFee, int beforePay, String cardType, String benefit1, String benefit1_detail, String benefit2, String benefit2_detail, String benefit3, String benefit3_detail, String cardLink,int count) {
+    public CardEntity(Long cardCode, String cardName, String cardPhoto, String cardCompany, int annualFee, int beforePay, String cardType, String benefit1, String benefit1_detail, String benefit2, String benefit2_detail, String benefit3, String benefit3_detail, String cardLink,int count) {
         this.cardCode = cardCode;
         this.cardName = cardName;
         this.cardPhoto = cardPhoto;
@@ -91,7 +91,6 @@ public class CardEntity {
         this.benefit3 = benefit3;
         this.benefit3_detail = benefit3_detail;
         this.cardLink = cardLink;
-
         return this;
     }
